Resolve "Data Source Deletion Warning/Force"

Closes #36

See merge request b650/Deep-Lynx!24

+++ src/data_access_layer/repositories/data_warehouse/import/data_source_repository.ts
@@ -1,11 +1,12 @@
-import RepositoryInterface, { QueryOptions, Repository } from '../../repository';
-import DataSourceRecord, { DataSource } from '../../../../data_warehouse/import/data_source';
+import RepositoryInterface, {QueryOptions, Repository} from '../../repository';
+import DataSourceRecord, {DataSource} from '../../../../data_warehouse/import/data_source';
 import DataSourceMapper from '../../../mappers/data_warehouse/import/data_source_mapper';
 import HttpDataSourceImpl from '../../../../data_warehouse/import/http_data_source_impl';
 import StandardDataSourceImpl from '../../../../data_warehouse/import/standard_data_source_impl';
 import Result from '../../../../common_classes/result';
-import { User } from '../../../../access_management/user';
-import { PoolClient } from 'pg';
+import {User} from '../../../../access_management/user';
+import {PoolClient} from 'pg';
+import ImportMapper from '../../../mappers/data_warehouse/import/import_mapper';
 
 /*
     DataSourceRepository contains methods for persisting and retrieving data sources
@@ -20,11 +21,22 @@ export default class DataSourceRepository extends Repository implements Reposito
     #mapper = DataSourceMapper.Instance;
     #factory = new DataSourceFactory();
 
-    delete(t: DataSource): Promise<Result<boolean>> {
+    async delete(t: DataSource, forceDelete?: boolean): Promise<Result<boolean>> {
         if (!t.DataSourceRecord || !t.DataSourceRecord.id)
             return Promise.resolve(Result.Failure(`cannot delete data source: no data source record or record lacking id`));
 
-        return this.#mapper.Delete(t.DataSourceRecord.id);
+        const hasImports = await ImportMapper.Instance.ExistForDataSource(t.DataSourceRecord.id);
+
+        if (!hasImports || forceDelete) return this.#mapper.Delete(t.DataSourceRecord.id);
+
+        return Promise.resolve(Result.Failure(`Data Source has data associated with it, this data must be removed or user must force delete.`));
+    }
+
+    async archive(u: User, t: DataSource): Promise<Result<boolean>> {
+        if (!t.DataSourceRecord || !t.DataSourceRecord.id)
+            return Promise.resolve(Result.Failure(`cannot archive data source: no data source record or record lacking id`));
+
+        return this.#mapper.Archive(t.DataSourceRecord.id, u.id!);
     }
 
     async findByID(id: string): Promise<Result<DataSource>> {
@@ -38,7 +50,7 @@ export default class DataSourceRepository extends Repository implements Reposito
     }
 
     async save(t: DataSource, user: User): Promise<Result<boolean>> {
-        if (!t.DataSourceRecord) return Promise.resolve(Result.Failure(`DataSource must have a data source record instantiated`));
+        if (!t.DataSourceRecord) return Promise.resolve(Result.Failure(`Data Source must have a data source record instantiated`));
 
         const errors = await t.DataSourceRecord.validationErrors();
         if (errors) return Promise.resolve(Result.Failure(`attached data source record does not pass validation ${errors.join(',')}`));
@@ -107,6 +119,11 @@ export default class DataSourceRepository extends Repository implements Reposito
         return this;
     }
 
+    archived(value: boolean) {
+        super.query('archived', 'eq', value);
+        return this;
+    }
+
     inactive() {
         super.query('active', 'eq', false);
         return this;
@@ -115,7 +132,7 @@ export default class DataSourceRepository extends Repository implements Reposito
     async list(options?: QueryOptions, transaction?: PoolClient): Promise<Result<(DataSource | undefined)[]>> {
         const results = await super.findAll<DataSourceRecord>(options, {
             transaction,
-            resultClass: DataSourceRecord
+            resultClass: DataSourceRecord,
         });
         if (results.isError) return Promise.resolve(Result.Pass(results));
